requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
markdownify==0.11.6
python-slugify==4.0.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
text-unidecode==1.3
urllib3==2.0.7
